Improve the shutdown procedure and shutdown logging strategy
The ClientBalancer sometimes enters the waiting phase of one minute,
just after the test has finished. This means the user must wait an
entire minute before the results are available, which is annoying.

This commit fixes the issue by explicitly calling the interrupt method
on the thread, so that it immediately begins the shutdown process. The
logging has also been updated to more accurately show the state of the
shutdown.

Furthermore, TPCHWorkers would sometimes enter an infinite loop, due to
the global workloadstate receiving the unhandled case EXIT. The bug is
likely caused by the AtomicInteger notDoneCount not being incremented
when a new TPCHWorker is added to the system by the ClientBalancer. This
commit is a convenient easy work-around to fixing the issue.

Fix #28